[refactor/#303] ChatSocketClient에서 ChatSocketService로 변경
- 에러 관리 책임을 도메인으로 이전
- data 계층 로직 단순화
