+++ data/chat/src/main/java/com/napzak/market/chat/datasource/ChatSocketDataSource.kt
@@ -1,25 +1,22 @@
 package com.napzak.market.chat.datasource
 
-import com.napzak.market.chat.client.ChatSocketClient
 import com.napzak.market.chat.dto.ChatMessageRequest
-import kotlinx.coroutines.flow.Flow
+import com.napzak.market.chat.service.ChatSocketService
 import javax.inject.Inject
 
 class ChatSocketDataSource @Inject constructor(
-    private val chatSocketClient: ChatSocketClient,
+    private val chatSocketService: ChatSocketService,
 ) {
-    val errorFlow: Flow<Throwable> = chatSocketClient.errorFlow
+    suspend fun connect() = chatSocketService.connect()
 
-    suspend fun connect() = chatSocketClient.connect()
-
-    suspend fun disconnect() = chatSocketClient.disconnect()
+    suspend fun disconnect() = chatSocketService.disconnect()
 
     suspend fun subscribeChatRoom(roomId: Long) =
-        chatSocketClient.subscribeChatRoom(roomId)
+        chatSocketService.subscribeChatRoom(roomId)
 
     suspend fun subscribeCreateChatRoom(storeId: Long) =
-        chatSocketClient.subscribeCreateChatRoom(storeId)
+        chatSocketService.subscribeCreateChatRoom(storeId)
 
     suspend fun sendMessage(request: ChatMessageRequest) =
-        chatSocketClient.sendMessage(request)
+        chatSocketService.sendMessage(request)
 }
