+++ data/chat/src/main/java/com/napzak/market/chat/repositoryimpl/ChatSocketRepositoryImpl.kt
@@ -5,16 +5,12 @@ import com.napzak.market.chat.mapper.toDomain
 import com.napzak.market.chat.mapper.toRequest
 import com.napzak.market.chat.model.SendMessage
 import com.napzak.market.chat.repository.ChatSocketRepository
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
 
 class ChatSocketRepositoryImpl @Inject constructor(
     private val chatSocketDataSource: ChatSocketDataSource,
 ) : ChatSocketRepository {
-
-    override val errorFlow: Flow<Throwable> = chatSocketDataSource.errorFlow
-
     override suspend fun connect() = runCatching {
         chatSocketDataSource.connect()
     }
