+++ domain/chat/src/main/java/com/napzak/market/chat/repository/ChatSocketRepository.kt
@@ -5,7 +5,6 @@ import com.napzak.market.chat.model.SendMessage
 import kotlinx.coroutines.flow.Flow
 
 interface ChatSocketRepository {
-    val errorFlow: Flow<Throwable>
     suspend fun connect(): Result<Unit>
     suspend fun disconnect(): Result<Unit>
     suspend fun subscribeChatRoom(roomId: Long, storeId: Long): Result<Flow<ReceiveMessage<*>>>
